Move warp and key states to data
