+++ src/main/kotlin/xyz/atrius/waystones/event/InfoEvent.kt
@@ -6,6 +6,7 @@ import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
 import org.bukkit.event.block.Action
 import org.bukkit.event.player.PlayerInteractEvent
+import xyz.atrius.waystones.data.WarpState
 import xyz.atrius.waystones.service.WarpNameService
 import xyz.atrius.waystones.utility.*
